add cterm values in the generated vim colorscheme
Several parts of the UI don't seem to work well with just the GUI
colors.
I think it is probably because of GUI colors being set to NONE.
Then Vim falls back to the existing base cterm color,
but I don't want that.
Adding valid cterm colors also makes the colorscheme more portable which
is a win.