fix(autocommands): ignore error in theme autorefresh
The error I encountered was when modifying the fugitive staged buffer.
In this case, apath matches theme.lua but the actual full path does not
exist which caused dofile to fail.

I think the only way this could cause a problem is when I'll eventually
rename the file, forget about the autocommand, and the autorefresh will
then fail silently.